requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
pillow
yt-dlp
